+++ spv/scripts_only.go
@@ -8,7 +8,7 @@ import (
 
 type GullibleHeadersClient struct{}
 
-func (g *GullibleHeadersClient) IsValidRootForHeight(merkleRoot *chainhash.Hash, height uint32) (bool, error) {
+func (g *GullibleHeadersClient) IsValidRootForHeight(ctx context.Context, merkleRoot *chainhash.Hash, height uint32) (bool, error) {
 	// DO NOT USE IN A REAL PROJECT due to security risks of accepting any merkle root as valid without verification
 	return true, nil
 }