+++ transaction/chaintracker/whatsonchain_test.go
@@ -53,7 +53,8 @@ func TestWhatsOnChainGetBlockHeaderSuccess(t *testing.T) {
 		client:  ts.Client(),
 	}
 
-	header, err := woc.GetBlockHeader(100)
+	ctx := t.Context()
+	header, err := woc.GetBlockHeader(ctx, 100)
 	if err != nil {
 		t.Fatalf("expected no error, got %v", err)
 		return // Add this return statement
@@ -81,7 +82,8 @@ func TestWhatsOnChainGetBlockHeaderNotFound(t *testing.T) {
 		client:  ts.Client(),
 	}
 
-	header, err := woc.GetBlockHeader(100)
+	ctx := t.Context()
+	header, err := woc.GetBlockHeader(ctx, 100)
 	if err != nil {
 		t.Fatalf("expected no error, got %v", err)
 	}
@@ -105,7 +107,8 @@ func TestWhatsOnChainGetBlockHeaderErrorResponse(t *testing.T) {
 		client:  ts.Client(),
 	}
 
-	header, err := woc.GetBlockHeader(100)
+	ctx := t.Context()
+	header, err := woc.GetBlockHeader(ctx, 100)
 	if err == nil {
 		t.Fatalf("expected error, got nil")
 	}
@@ -136,7 +139,8 @@ func TestWhatsOnChainIsValidRootForHeightSuccess(t *testing.T) {
 		client:  ts.Client(),
 	}
 
-	isValid, err := woc.IsValidRootForHeight(&merkleRootHash, 100)
+	ctx := t.Context()
+	isValid, err := woc.IsValidRootForHeight(ctx, &merkleRootHash, 100)
 	if err != nil {
 		t.Fatalf("expected no error, got %v", err)
 	}
@@ -168,7 +172,8 @@ func TestWhatsOnChainIsValidRootForHeightInvalidRoot(t *testing.T) {
 		client:  ts.Client(),
 	}
 
-	isValid, err := woc.IsValidRootForHeight(&differentMerkleRootHash, 100)
+	ctx := t.Context()
+	isValid, err := woc.IsValidRootForHeight(ctx, &differentMerkleRootHash, 100)
 	if err != nil {
 		t.Fatalf("expected no error, got %v", err)
 	}