+++ transaction/merklepath_test.go
@@ -97,7 +97,7 @@ func TestMerklePathComputeRootHex(t *testing.T) {
 type MyChainTracker struct{}
 
 // Implement the IsValidRootForHeight method on MyChainTracker.
-func (mct MyChainTracker) IsValidRootForHeight(root *chainhash.Hash, height uint32) (bool, error) {
+func (mct MyChainTracker) IsValidRootForHeight(ctx context.Context, root *chainhash.Hash, height uint32) (bool, error) {
 	// Convert BRC74Root hex string to a byte slice for comparison
 	// expectedRoot, _ := hex.DecodeString(BRC74Root)
 
@@ -118,7 +118,8 @@ func TestMerklePath_Verify(t *testing.T) {
 			Path:        BRC74JSON.Path,
 		}
 		tracker := MyChainTracker{}
-		result, err := path.VerifyHex(BRC74TXID1, tracker)
+		ctx := t.Context()
+		result, err := path.VerifyHex(ctx, BRC74TXID1, tracker)
 		require.NoError(t, err)
 		require.True(t, result)
 	})